Create bootstrap-navtabs-4-tabs
This Bootstrap code was designed to allow for keyboard selectable navigational tabs. This example uses four tabs. Each tab's div can be copied or deleted depending on your needs.

This code is solely intended to work as an article template solely compatible with TeamDynamix and is not intended for other platforms as is.